+++ core/src/test/scala/org/apache/pekko/persistence/jdbc/journal/dao/legacy/TagsSerializationTest.scala
@@ -3,10 +3,9 @@
  * Copyright (C) 2019 - 2021 Lightbend Inc. <https://www.lightbend.com>
  */
 
-package org.apache.pekko.persistence.jdbc.journal.dao
+package org.apache.pekko.persistence.jdbc.journal.dao.legacy
 
 import org.apache.pekko.persistence.jdbc.SharedActorSystemTestSpec
-import org.apache.pekko.persistence.jdbc.journal.dao.legacy.{ decodeTags, encodeTags }
 
 class TagsSerializationTest extends SharedActorSystemTestSpec {
   "Encode" should "no tags" in {
